Feat : ErrorMessage 이넘에 상수 수정
ErrorMessage 이넘에 불필요한 상수 삭제
