+++ BE/src/main/java/com/example/todolist/exception/ErrorMessage.java
@@ -1,0 +1,18 @@
+package com.example.todolist.exception;
+
+public enum ErrorMessage {
+    ENTITY_NOT_CREATE("해당 엔티티가 만들어지지 않았습니다."),
+    INVALID_USER("접근 권한이 없는 유저입니다."),
+    DUPLICATED_ID("이미 사용중인 아이디입니다."),
+    LOGIN_FAILED("아이디 또는 비밀번호가 틀립니다. 다시 로그인 해주세요.");
+
+    private final String errorMessage;
+
+    ErrorMessage(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+}
